float unit_price for tickets

+++ db/migrate/20130307143700_create_tickets.rb
@@ -2,7 +2,7 @@ class CreateTickets < ActiveRecord::Migration
   def change
     create_table :tickets do |t|
       t.string :name
-      t.decimal :unit_price
+      t.float :unit_price
 
       t.timestamps
     end
